# local-shared-data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/block_data/": {
      "get": {
        "tags": [
          "block_data"
        ],
        "summary": "Read Block Data",
        "operationId": "read_block_data",
        "parameters": [
          {
            "name": "block_timestamp_start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Block Timestamp Start"
            }
          },
          {
            "name": "block_timestamp_end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Block Timestamp End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlockData"
                  },
                  "title": "Response Block Data-Read Block Data"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exchange_rate/": {
      "get": {
        "tags": [
          "exchange_rate"
        ],
        "summary": "Read Exchange Rate",
        "operationId": "read_exchange_rate",
        "parameters": [
          {
            "name": "exchange_rate_timestamp_start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Exchange Rate Timestamp Start"
            }
          },
          {
            "name": "exchange_rate_timestamp_end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Exchange Rate Timestamp End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExchangeRate"
                  },
                  "title": "Response Exchange Rate-Read Exchange Rate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mwh_revenue/": {
      "get": {
        "tags": [
          "mwh_revenue"
        ],
        "summary": "Get Mwh Revenue",
        "operationId": "get_mwh_revenue",
        "parameters": [
          {
            "name": "timestamp_start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Timestamp Start"
            }
          },
          {
            "name": "timestamp_end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Timestamp End"
            }
          },
          {
            "name": "asic_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MWHRevenue"
                  },
                  "title": "Response Mwh Revenue-Get Mwh Revenue"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/asic/": {
      "get": {
        "tags": [
          "asic"
        ],
        "summary": "List Asics",
        "operationId": "list_asics",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ASIC"
                  },
                  "type": "array",
                  "title": "Response Asic-List Asics"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ASIC": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "asic_slug": {
            "type": "string",
            "title": "Asic Slug"
          },
          "asic_name": {
            "type": "string",
            "title": "Asic Name"
          },
          "asic_hash_rate": {
            "type": "number",
            "title": "Asic Hash Rate"
          },
          "asic_power": {
            "type": "number",
            "title": "Asic Power"
          }
        },
        "type": "object",
        "required": [
          "id",
          "asic_slug",
          "asic_name",
          "asic_hash_rate",
          "asic_power"
        ],
        "title": "ASIC"
      },
      "BlockData": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "block_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Block Timestamp"
          },
          "block_number": {
            "type": "integer",
            "title": "Block Number"
          },
          "block_subsidy": {
            "type": "number",
            "title": "Block Subsidy"
          },
          "block_transaction_fees": {
            "type": "number",
            "title": "Block Transaction Fees"
          },
          "network_hash_rate": {
            "type": "number",
            "title": "Network Hash Rate"
          }
        },
        "type": "object",
        "required": [
          "id",
          "block_timestamp",
          "block_number",
          "block_subsidy",
          "block_transaction_fees",
          "network_hash_rate"
        ],
        "title": "BlockData"
      },
      "ExchangeRate": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "block_number": {
            "type": "integer",
            "title": "Block Number"
          },
          "exchange_rate": {
            "type": "number",
            "title": "Exchange Rate"
          },
          "exchange_rate_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Exchange Rate Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "block_number",
          "exchange_rate",
          "exchange_rate_timestamp"
        ],
        "title": "ExchangeRate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MWHRevenue": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "asic_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asic Id"
          },
          "mwh_btc_revenue": {
            "type": "number",
            "title": "Mwh Btc Revenue"
          },
          "mwh_usd_revenue": {
            "type": "number",
            "title": "Mwh Usd Revenue"
          },
          "mwh_revenue_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Mwh Revenue Timestamp"
          },
          "block_number": {
            "type": "integer",
            "title": "Block Number"
          }
        },
        "type": "object",
        "required": [
          "id",
          "asic_id",
          "mwh_btc_revenue",
          "mwh_usd_revenue",
          "mwh_revenue_timestamp",
          "block_number"
        ],
        "title": "MWHRevenue"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}